address code review feedback: encapsulate the fields of CacheEntryTokens
